better comments, some refactor
