+++ src/main.h
@@ -5,7 +5,5 @@
 #ifndef CUPIDFM_MAIN_H
 #define CUPIDFM_MAIN_H
 
-void path_join(char *result, const char *path1, const char *path2);
-
 
 #endif //CUPIDFM_MAIN_H
