+++ src/utils.c
@@ -16,8 +16,13 @@
 
 #define MAX_PATH_LENGTH 256
 #define MAX_DISPLAY_LENGTH 32
-
-
+/**
+ * Function to join two paths together
+ *
+ * @param result the resulting path
+ * @param base the base path
+ * @param extra the extra path to append
+ */
 void die(int r, const char *format, ...) {
 	fprintf(stderr, "The program used die()\n");
 	fprintf(stderr, "The last errno was %d/%s\n", errno, strerror(errno));
@@ -33,15 +38,26 @@ void die(int r, const char *format, ...) {
 
 	exit(r);
 }
-
-
+/**
+ * Function to join two paths together
+ *
+ * @param result the resulting path
+ * @param base the base path
+ * @param extra the extra path to append
+ */
 void create_file(const char *filename) {
     FILE *f = fopen(filename, "w");
     if (f == NULL)
         die(1, "Couldn't create file %s", filename);
     fclose(f);
 }
-
+/**
+ * Function to join two paths together
+ *
+ * @param result the resulting path
+ * @param base the base path
+ * @param extra the extra path to append
+ */
 void browse_files(const char *directory) {
     char command[256];
     snprintf(command, sizeof(command), "xdg-open %s", directory);
@@ -56,7 +72,11 @@ void browse_files(const char *directory) {
         printf("Error: The file manager returned a non-zero status.\n");
     }
 }
-
+/**
+ * Function to display the contents of a directory in the terminal
+ *
+ * @param directory the directory to display
+ */
 void display_files(const char *directory) {
     DIR *dir;
     struct dirent *entry;
@@ -72,7 +92,11 @@ void display_files(const char *directory) {
 
     closedir(dir);
 }
-
+/**
+ * Function to preview the contents of a file in the terminal
+ *
+ * @param filename the name of the file to preview
+ */
 void preview_file(const char *filename) {
     FILE *file = fopen(filename, "r");
     if (file == NULL) {
@@ -109,7 +133,16 @@ void preview_file(const char *filename) {
     endwin();
     fclose(file);
 }
-
+/**
+ * Function to change the current directory and update the list of files
+ *
+ * @param new_directory the new directory to change to
+ * @param files the list of files in the directory
+ * @param num_files the number of files in the directory
+ * @param selected_entry the index of the selected entry
+ * @param start_entry the index of the first entry displayed
+ * @param end_entry the index of the last entry displayed
+ */
 bool is_directory(const char *path, const char *filename) {
     struct stat path_stat;
     char full_path[MAX_PATH_LENGTH];
@@ -121,3 +154,27 @@ bool is_directory(const char *path, const char *filename) {
     return false; // Correct: Do not assume it's a directory if stat fails
 }
 
+/** Function to join two paths
+ *
+ * @param result the resulting path
+ * @param base the base path
+ * @param extra the extra path
+ */
+void path_join(char *result, const char *base, const char *extra) {
+    size_t base_len = strlen(base);
+    size_t extra_len = strlen(extra);
+
+    if (base_len == 0) {
+        strncpy(result, extra, MAX_PATH_LENGTH);
+    } else if (extra_len == 0) {
+        strncpy(result, base, MAX_PATH_LENGTH);
+    } else {
+        if (base[base_len - 1] == '/') {
+            snprintf(result, MAX_PATH_LENGTH, "%s%s", base, extra);
+        } else {
+            snprintf(result, MAX_PATH_LENGTH, "%s/%s", base, extra);
+        }
+    }
+
+    result[MAX_PATH_LENGTH - 1] = '\0';
+}
