+++ src/utils.h
@@ -11,9 +11,9 @@
 __attribute__((noreturn))
 void die(int r, const char *format, ...);
 
-void path_join(char *result, const char *base, const char *extra);
 void create_file(const char *filename);
 void display_files(const char *directory);
 void preview_file(const char *filename);
 void change_directory(const char *new_directory, const char ***files, int *num_files, int *selected_entry, int *start_entry, int *end_entry);
 bool is_directory(const char *path, const char *filename);
+void path_join(char *result, const char *base, const char *extra);
